pyproject.toml
[project]
name = "blarp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "moderngl[headless]>=5.12.0",
    "numpy>=2.3.4",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.3",
]
